Default Pensum in Academic Year Details

+++ Mhotivo/Models/AcademicYearDetailsModel.cs
@@ -59,7 +59,7 @@ namespace Mhotivo.Models
         public Meister Teacher { get; set; }
 
         [Display(Name = "Id Año Academico")]
-        public long AcademicYearId { get; set; }
+        public int AcademicYearId { get; set; }
 
     }
